feat(sauvegarde): chiffrer les archives, les sortir du serveur, et cesser d'exposer le mot de passe des bases
En preparant la page Securite du site, il fallait repondre par ecrit a
« vos sauvegardes sont-elles chiffrees ? ». La reponse etait non. En
regardant de pres, ce n'etait pas le plus grave.

## Le mot de passe des bases etait lisible par tout le serveur

    mysqldump -h %s -P %d -u %s -p%s %s | gzip > %s

Le mot de passe passait par la ligne de commande. Sur un hebergement
mutualise, `ps aux` montre la ligne de commande de TOUS les processus :
n'importe quel autre compte du serveur pouvait lire le mot de passe de la
base d'une ecole pendant la duree du dump. Chaque nuit a 2 h, une fenetre
par etablissement.

Il passe desormais par `--defaults-extra-file`, un fichier en 0600 efface
dans un `finally` — sans quoi un dump qui echoue laisserait le mot de passe
dans /tmp.

L'hote et le port n'etaient pas echappes non plus. Ils viennent d'une source
de confiance, donc le risque etait faible ; « faible parce que la source est
de confiance » est le raisonnement qui casse le jour ou la source change.

## Les archives sont chiffrees, et deposees ailleurs

AES-256-CBC avec `-pbkdf2` (sans lui, OpenSSL derive encore la cle par un MD5
a une iteration). La cle est lue depuis un fichier, jamais depuis la ligne de
commande : c'eut ete rejouer le defaut qu'on corrige.

`SAUVEGARDE_DISQUE_HORS_SITE` depose une copie hors du serveur sauvegarde. Une
archive qui vit a cote de la base qu'elle protege ne protege que d'une fausse
manipulation, et disparait avec le serveur.

**La cle manquante n'interrompt pas la sauvegarde.** Une nuit sans sauvegarde
coute plus cher qu'une sauvegarde en clair. L'archive est prise, la ligne
enregistre `est_chiffre = false`, et la commande le dit fort. Ce qui manque
devient mesurable au lieu d'etre suppose.

## Un defaut evite de justesse

J'avais d'abord ecrit `set -o pipefail;` en tete du tuyau. `exec()` de PHP
passe par `/bin/sh`, qui est `dash` : « Illegal option -o pipefail », commande
jamais lancee. Cela aurait arrete TOUTES les sauvegardes, chaque nuit, en
silence — decouvert le jour d'une restauration. Le test l'a vu parce qu'il
execute la commande au lieu de lire la chaine. On appelle donc bash
explicitement.

`pipefail` reste indispensable : sans lui, un `mysqldump` qui echoue suivi
d'un `gzip` qui reussit rend 0, et l'on enregistre comme reussie une archive
vide. Un second filet verifie l'archive produite (`gzip -t`, taille non nulle).

## Verifications

19, dont l'aller-retour complet sur disque : chiffrer une arborescence reelle,
verifier que le nom de l'etudiant ne s'y lit plus, la dechiffrer, et retrouver
l'octet pres ce qu'on avait donne. Plus le contraste : sans cle, ce meme nom
se lit apres un simple `gunzip`. Plus deux preuves d'injection executees.

Suite complete : 333 tests, aucun echec (37 erreurs de base absente, propres a
ce conteneur).

## Au passage

`config/group_portal.php` promettait un entier et rendait une chaine des que
`.env` posait la variable — ce que `.env.example` fait. Le depot avait donc
deux comportements selon qu'on suivait ou non sa propre procedure
d'installation, et un seul passait les tests.

+++ config/group_portal.php
@@ -22,7 +22,13 @@ return [
     | Debounce applied client-side (Alpine) before the Livewire state is updated.
     | Tuned for quick clicks without triggering one HTTP roundtrip per keystroke.
     */
-    'period_selector_debounce_ms' => env('GROUP_PORTAL_PERIOD_DEBOUNCE_MS', 300),
+    // `(int)` a l'air superflu et ne l'est pas : `env()` rend TOUJOURS une
+    // chaine pour une valeur posee dans `.env`, et `.env.example` pose
+    // justement celle-ci. Le reglage valait donc 300 en entier sur une
+    // installation sans `.env`, et '300' en chaine des qu'on suit la
+    // procedure d'installation documentee — deux comportements pour un meme
+    // depot, dont un seul passait les tests.
+    'period_selector_debounce_ms' => (int) env('GROUP_PORTAL_PERIOD_DEBOUNCE_MS', 300),
 
     /*
     |--------------------------------------------------------------------------
